requirements.txt
Django==1.10.8
djangorestframework==3.7.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
lxml==4.5.0
pkg-resources==0.0.0
pytz==2019.3
soaplib==2.0.0b2
